fix: elect coordinator logic, ParseProduceRequest; minimal intergration test
