--- docker-compose.yml
@@ -1,25 +1,0 @@
-# docker-compose.yml
-services:
-  broker:
-    build: .
-    image: gokafk:dev
-    ports:
-      - "10000:10000"
-    healthcheck:
-      test: ["CMD-SHELL", "nc -z localhost 10000 || exit 1"]
-      interval: 2s
-      timeout: 5s
-      retries: 10
-      start_period: 3s
-
-  test-runner:
-    image: node:20-alpine
-    working_dir: /app
-    volumes:
-      - ./test/kafkajs:/app
-    environment:
-      - KAFKA_BROKER=broker:10000
-    depends_on:
-      broker:
-        condition: service_healthy
-    command: sh -c "npm install && npm test"
