+++ test/kafkajs/jest.config.js
@@ -2,8 +2,6 @@
 /** @type {import('jest').Config} */
 module.exports = {
   testEnvironment: 'node',
-  globalSetup: './globalSetup.js',
-  globalTeardown: './globalTeardown.js',
   testTimeout: 30000,
   verbose: true,
   reporters: ['default'],
